Split AuthorizationFeatureTests
